+++ discord/ui/view.py
@@ -260,7 +260,7 @@ class BaseView(ItemInterface):
             Maximum number of children has been exceeded
         """
 
-        if len(self.children) >= self.MAX_ITEMS:
+        if len(self) >= self.MAX_ITEMS:
             raise ValueError("maximum number of children exceeded")
 
         if not isinstance(item, ViewItem):
@@ -287,6 +287,8 @@ class BaseView(ItemInterface):
 
         if isinstance(item, (str, int)):
             item = self.get_item(item)
+            if not item:
+                return self
         try:
             if item.parent is self:
                 self.children.remove(item)
@@ -549,6 +551,21 @@ class BaseView(ItemInterface):
     def message(self, value):
         self._message = value
 
+    @classmethod
+    def from_message(
+        cls,
+        message: Message,
+        /,
+        *,
+        timeout: float | None = 180.0,
+        disable_on_timeout: bool = False,
+        store: bool = True,
+    ) -> BaseView:
+        view = cls(timeout=timeout, disable_on_timeout=disable_on_timeout, store=store)
+        for component in message.components:
+            view.add_item(_component_to_item(component))
+        return view
+
 
 class View(BaseView):
     """Represents a legacy UI view for V1 components :class:`~discord.ui.Button` and :class:`~discord.ui.Select`.
@@ -568,6 +585,8 @@ class View(BaseView):
     timeout: Optional[:class:`float`]
         Timeout in seconds from last interaction with the UI before no longer accepting input. Defaults to 180.0.
         If ``None`` then there is no timeout.
+    disable_on_timeout: :class:`bool`
+        Whether to disable the view when the timeout is reached. Defaults to ``False``.
     store: Optional[:class:`bool`]
         Whether this view should be stored for callback listening. Setting it to ``False`` will ignore item callbacks and prevent their values from being refreshed. Defaults to ``True``.
 
@@ -578,8 +597,6 @@ class View(BaseView):
         If ``None`` then there is no timeout.
     children: List[:class:`ViewItem`]
         The list of children attached to this view.
-    disable_on_timeout: :class:`bool`
-        Whether to disable the view when the timeout is reached. Defaults to ``False``.
     message: Optional[:class:`.Message`]
         The message that this view is attached to.
         If ``None`` then the view has not been sent with a message.
@@ -650,7 +667,13 @@ class View(BaseView):
 
     @classmethod
     def from_message(
-        cls, message: Message, /, *, timeout: float | None = 180.0
+        cls,
+        message: Message,
+        /,
+        *,
+        timeout: float | None = 180.0,
+        disable_on_timeout: bool = False,
+        store: bool = True,
     ) -> View:
         """Converts a message's components into a :class:`View`.
 
@@ -665,6 +688,10 @@ class View(BaseView):
             The message with components to convert into a view.
         timeout: Optional[:class:`float`]
             The timeout of the converted view.
+        disable_on_timeout: :class:`bool`
+            Whether to disable the view when the timeout is reached. Defaults to ``False``.
+        store: Optional[:class:`bool`]
+            Whether this view should be stored for callback listening. Setting it to ``False`` will ignore item callbacks and prevent their values from being refreshed. Defaults to ``True``.
 
         Returns
         -------
@@ -672,7 +699,7 @@ class View(BaseView):
             The converted view. This always returns a :class:`View` and not
             one of its subclasses.
         """
-        view = View(timeout=timeout)
+        view = View(timeout=timeout, disable_on_timeout=disable_on_timeout, store=store)
         for component in _walk_all_components(message.components):
             view.add_item(_component_to_item(component))
         return view
@@ -684,6 +711,8 @@ class View(BaseView):
         /,
         *,
         timeout: float | None = 180.0,
+        disable_on_timeout: bool = False,
+        store: bool = True,
     ) -> View:
         """Converts a list of component dicts into a :class:`View`.
 
@@ -693,6 +722,10 @@ class View(BaseView):
             The list of components to convert into a view.
         timeout: Optional[:class:`float`]
             The timeout of the converted view.
+        disable_on_timeout: :class:`bool`
+            Whether to disable the view when the timeout is reached. Defaults to ``False``.
+        store: Optional[:class:`bool`]
+            Whether this view should be stored for callback listening. Setting it to ``False`` will ignore item callbacks and prevent their values from being refreshed. Defaults to ``True``.
 
         Returns
         -------
@@ -700,7 +733,7 @@ class View(BaseView):
             The converted view. This always returns a :class:`View` and not
             one of its subclasses.
         """
-        view = View(timeout=timeout)
+        view = View(timeout=timeout, disable_on_timeout=disable_on_timeout, store=store)
         components = [_component_factory(d) for d in data]
         for component in _walk_all_components(components):
             view.add_item(_component_to_item(component))
@@ -807,6 +840,8 @@ class DesignerView(BaseView):
     timeout: Optional[:class:`float`]
         Timeout in seconds from last interaction with the UI before no longer accepting input. Defaults to 180.0.
         If ``None`` then there is no timeout.
+    disable_on_timeout: :class:`bool`
+        Whether to disable the view's items when the timeout is reached. Defaults to ``False``.
     store: Optional[:class:`bool`]
         Whether this view should be stored for callback listening. Setting it to ``False`` will ignore item callbacks and prevent their values from being refreshed. Defaults to ``True``.
 
@@ -817,8 +852,6 @@ class DesignerView(BaseView):
         If ``None`` then there is no timeout.
     children: List[:class:`ViewItem`]
         The list of items attached to this view.
-    disable_on_timeout: :class:`bool`
-        Whether to disable the view's items when the timeout is reached. Defaults to ``False``.
     message: Optional[:class:`.Message`]
         The message that this view is attached to.
         If ``None`` then the view has not been sent with a message.
@@ -858,8 +891,14 @@ class DesignerView(BaseView):
 
     @classmethod
     def from_message(
-        cls, message: Message, /, *, timeout: float | None = 180.0
-    ) -> View:
+        cls,
+        message: Message,
+        /,
+        *,
+        timeout: float | None = 180.0,
+        disable_on_timeout: bool = False,
+        store: bool = True,
+    ) -> DesignerView:
         """Converts a message's components into a :class:`DesignerView`.
 
         The :attr:`.Message.components` of a message are read-only
@@ -873,14 +912,20 @@ class DesignerView(BaseView):
             The message with components to convert into a view.
         timeout: Optional[:class:`float`]
             The timeout of the converted view.
+        disable_on_timeout: :class:`bool`
+            Whether to disable the view when the timeout is reached. Defaults to ``False``.
+        store: Optional[:class:`bool`]
+            Whether this view should be stored for callback listening. Setting it to ``False`` will ignore item callbacks and prevent their values from being refreshed. Defaults to ``True``.
 
         Returns
         -------
         :class:`View`
-            The converted view. This always returns a :class:`View` and not
+            The converted view. This always returns a :class:`DesignerView` and not
             one of its subclasses.
         """
-        view = DesignerView(timeout=timeout)
+        view = DesignerView(
+            timeout=timeout, disable_on_timeout=disable_on_timeout, store=store
+        )
         for component in message.components:
             view.add_item(_component_to_item(component))
         return view
@@ -892,7 +937,9 @@ class DesignerView(BaseView):
         /,
         *,
         timeout: float | None = 180.0,
-    ) -> View:
+        disable_on_timeout: bool = False,
+        store: bool = True,
+    ) -> DesignerView:
         """Converts a list of component dicts into a :class:`DesignerView`.
 
         Parameters
@@ -901,42 +948,311 @@ class DesignerView(BaseView):
             The list of components to convert into a view.
         timeout: Optional[:class:`float`]
             The timeout of the converted view.
+        disable_on_timeout: :class:`bool`
+            Whether to disable the view when the timeout is reached. Defaults to ``False``.
+        store: Optional[:class:`bool`]
+            Whether this view should be stored for callback listening. Setting it to ``False`` will ignore item callbacks and prevent their values from being refreshed. Defaults to ``True``.
 
         Returns
         -------
         :class:`DesignerView`
-            The converted view. This always returns a :class:`View` and not
+            The converted view. This always returns a :class:`DesignerView` and not
             one of its subclasses.
         """
-        view = DesignerView(timeout=timeout)
+        view = DesignerView(
+            timeout=timeout, disable_on_timeout=disable_on_timeout, store=store
+        )
         components = [_component_factory(d) for d in data]
         for component in components:
             view.add_item(_component_to_item(component))
         return view
 
-    def add_item(self, item: ViewItem[V]) -> Self:
+    def add_item(
+        self,
+        item: ViewItem[V],
+        *,
+        index: int | None = None,
+        before: ViewItem[V] | str | int | None = None,
+        after: ViewItem[V] | str | int | None = None,
+        into: ViewItem[V] | str | int | None = None,
+    ) -> Self:
         """Adds an item to the view.
+
+        .. warning::
+
+            You may specify only **one** of ``index``, ``before``, & ``after``. ``into`` will work together with those parameters.
+
+        .. versionchanged:: 2.7.1
+            Added new parameters ``index``, ``before``, ``after``, & ``into``.
 
         Parameters
         ----------
         item: :class:`ViewItem`
             The item to add to the view.
+        index: Optional[class:`int`]
+            Add the new item at the specific index of :attr:`children`. Same behavior as Python's :func:`~list.insert`.
+        before: Optional[Union[:class:`ViewItem`, :class:`int`, :class:`str`]]
+            Add the new item **before** the specified item. If an :class:`int` is provided, the item will be detected by ``id``, otherwise by ``custom_id``.
+        after: Optional[Union[:class:`ViewItem`, :class:`int`, :class:`str`]]
+            Add the new item **after** the specified item. If an :class:`int` is provided, the item will be detected by ``id``, otherwise by ``custom_id``.
+        into: Optional[Union[:class:`ViewItem`, :class:`int`, :class:`str`]]
+            Add the new item **into** the specified item. This would be equivalent to `into.add_item(item)`, where `into` is a :class:`ViewItem`.
+            If an :class:`int` is provided, the item will be detected by ``id``, otherwise by ``custom_id``.
 
         Raises
         ------
         TypeError
-            An :class:`ViewItem` was not passed.
+            A :class:`ViewItem` was not passed.
         ValueError
-            Maximum number of items has been exceeded (40)
+            Maximum number of items has been exceeded (40),
+            or a searched item could not be found in the view.
         """
+        if sum(x is not None for x in (before, after, index)) > 1:
+            raise ValueError("Can only specify one of before, after, and index.")
 
-        if isinstance(item._underlying, (SelectComponent, ButtonComponent)):
+        if len(self) + len(item) > self.MAX_ITEMS:
+            raise ValueError("maximum number of children exceeded (40)")
+
+        if not isinstance(item, ViewItem):
+            raise TypeError(f"expected item to be ViewItem, not {item.__class__!r}")
+
+        if isinstance(item.underlying, (SelectComponent, ButtonComponent)):
             raise ValueError(
                 f"cannot add Select or Button to DesignerView directly. Use ActionRow instead."
             )
+        if into and isinstance(into, (str, int)):
+            parent = self.get_item(into)
+            if not parent:
+                raise ValueError(f"Could not find {into} in view.")
+        else:
+            parent = into or self
 
-        super().add_item(item)
+        if before is not None or after is not None:
+            ref = before or after or 0
+            if isinstance(ref, (int, str)):
+                ref = parent.get_item(ref)
+            try:
+                if ref.parent is parent:
+                    i = parent.items.index(ref)
+                    item.parent = parent
+                    if before:
+                        parent.items.insert(i, item)
+                    else:
+                        parent.items.insert(i + 1, item)
+                else:
+                    if isinstance(ref.parent.underlying, ContainerComponent):
+                        ref.parent.add_item(item, before=before, after=after, into=into)
+                    else:
+                        ref.parent.add_item(item, before=before, after=after)
+            except (ValueError, AttributeError):
+                raise ValueError(f"Could not find {before or after} in view.")
+            return self
+
+        elif index is not None:
+            item.parent = parent
+            parent.items.insert(index, item)
+            return self
+
+        item.parent = parent
+        parent.items.append(item)
         return self
+
+    def replace_item(
+        self, original_item: ViewItem[V] | str | int, new_item: ViewItem[V]
+    ) -> Self:
+        """Directly replace an item in this view.
+        If an :class:`int` is provided, the item will be replaced by ``id``, otherwise by ``custom_id``.
+
+        Parameters
+        ----------
+        original_item: Union[:class:`ViewItem`, :class:`int`, :class:`str`]
+            The item, item ``id``, or item ``custom_id`` to replace in the view.
+        new_item: :class:`ViewItem`
+            The new item to insert into the view.
+
+        Returns
+        -------
+        :class:`BaseView`
+            The view instance.
+        """
+
+        if not original_item:
+            raise TypeError(
+                f"expected original_item to be a valid ViewItem, str, or int, not {new_item.__class__!r}"
+            )
+        if not isinstance(new_item, ViewItem):
+            raise TypeError(
+                f"expected new_item to be ViewItem, not {new_item.__class__!r}"
+            )
+
+        if isinstance(original_item, (str, int)):
+            original_item = self.get_item(original_item)
+        if not original_item:
+            raise ValueError(f"Could not find {original_item} in view.")
+        try:
+            if original_item.parent is self:
+                i = self.children.index(original_item)
+                new_item.parent = self
+                self.children[i] = new_item
+                original_item.parent = None
+            else:
+                original_item.parent.replace_item(original_item, new_item)
+        except ValueError:
+            raise ValueError(f"Could not find {original_item} in view.")
+        return self
+
+    def add_row(
+        self,
+        *items: ViewItem[V],
+        id: int | None = None,
+    ) -> Self:
+        """Adds an :class:`ActionRow` to the view.
+
+        To append a pre-existing :class:`ActionRow`, use :meth:`add_item` instead.
+
+        Parameters
+        ----------
+        *items: Union[:class:`Button`, :class:`Select`]
+            The items this action row contains.
+        id: Optiona[:class:`int`]
+            The action row's ID.
+        """
+        from .action_row import ActionRow
+
+        row = ActionRow(*items, id=id)
+
+        return self.add_item(row)
+
+    def add_container(
+        self,
+        *items: ViewItem[V],
+        id: int | None = None,
+    ) -> Self:
+        """Adds a :class:`Container` to the view.
+
+        To append a pre-existing :class:`Container`, use the
+        :meth:`add_item` method, instead.
+
+        Parameters
+        ----------
+        *items: :class:`ViewItem`
+            The items contained in this container.
+        id: Optional[:class:`int`]
+            The container's ID.
+        """
+        from .container import Container
+
+        container = Container(*items, id=id)
+
+        return self.add_item(container)
+
+    def add_section(
+        self,
+        *items: ViewItem[V],
+        accessory: ViewItem[V],
+        id: int | None = None,
+    ) -> Self:
+        """Adds a :class:`Section` to the view.
+
+        To append a pre-existing :class:`Section`, use the
+        :meth:`add_item` method, instead.
+
+        Parameters
+        ----------
+        *items: :class:`ViewItem`
+            The items contained in this section, up to 3.
+            Currently only supports :class:`~discord.ui.TextDisplay`.
+        accessory: Optional[:class:`ViewItem`]
+            The section's accessory. This is displayed in the top right of the section.
+            Currently only supports :class:`~discord.ui.Button` and :class:`~discord.ui.Thumbnail`.
+        id: Optional[:class:`int`]
+            The section's ID.
+        """
+        from .section import Section
+
+        section = Section(*items, accessory=accessory, id=id)
+
+        return self.add_item(section)
+
+    def add_text(self, content: str, id: int | None = None) -> Self:
+        """Adds a :class:`TextDisplay` to the view.
+
+        Parameters
+        ----------
+        content: :class:`str`
+            The content of the TextDisplay
+        id: Optiona[:class:`int`]
+            The text displays' ID.
+        """
+        from .text_display import TextDisplay
+
+        text = TextDisplay(content, id=id)
+
+        return self.add_item(text)
+
+    def add_gallery(
+        self,
+        *items: MediaGalleryItem,
+        id: int | None = None,
+    ) -> Self:
+        """Adds a :class:`MediaGallery` to the view.
+
+        To append a pre-existing :class:`MediaGallery`, use :meth:`add_item` instead.
+
+        Parameters
+        ----------
+        *items: :class:`MediaGalleryItem`
+            The media this gallery contains.
+        id: Optiona[:class:`int`]
+            The gallery's ID.
+        """
+        from .media_gallery import MediaGallery
+
+        g = MediaGallery(*items, id=id)
+
+        return self.add_item(g)
+
+    def add_file(self, url: str, spoiler: bool = False, id: int | None = None) -> Self:
+        """Adds a :class:`File` to the view.
+
+        Parameters
+        ----------
+        url: :class:`str`
+            The URL of this file's media. This must be an ``attachment://`` URL that references a :class:`~discord.File`.
+        spoiler: Optional[:class:`bool`]
+            Whether the file has the spoiler overlay. Defaults to ``False``.
+        id: Optiona[:class:`int`]
+            The file's ID.
+        """
+        from .file import File
+
+        f = File(url, spoiler=spoiler, id=id)
+
+        return self.add_item(f)
+
+    def add_separator(
+        self,
+        *,
+        divider: bool = True,
+        spacing: SeparatorSpacingSize = SeparatorSpacingSize.small,
+        id: int | None = None,
+    ) -> Self:
+        """Adds a :class:`Separator` to the container.
+
+        Parameters
+        ----------
+        divider: :class:`bool`
+            Whether the separator is a divider. Defaults to ``True``.
+        spacing: :class:`~discord.SeparatorSpacingSize`
+            The spacing size of the separator. Defaults to :attr:`~discord.SeparatorSpacingSize.small`.
+        id: Optional[:class:`int`]
+            The separator's ID.
+        """
+        from .separator import Separator
+
+        s = Separator(divider=divider, spacing=spacing, id=id)
+
+        return self.add_item(s)
 
     def _refresh(self, components: list[Component]):
         # Refreshes view data using discord's values
@@ -977,7 +1293,7 @@ class ViewStore:
         }
         return list(views.values())
 
-    def __verify_integrity(self):
+    def __verify_integrity(self) -> None:
         to_remove: list[tuple[int, int | None, str]] = []
         for k, (view, _) in self._views.items():
             if view.is_finished():
@@ -986,7 +1302,7 @@ class ViewStore:
         for k in to_remove:
             del self._views[k]
 
-    def add_view(self, view: BaseView, message_id: int | None = None):
+    def add_view(self, view: BaseView, message_id: int | None = None) -> None:
         if not view._store:
             return
         self.__verify_integrity()
@@ -1002,7 +1318,7 @@ class ViewStore:
         if message_id is not None:
             self._synced_message_views[message_id] = view
 
-    def remove_view(self, view: BaseView):
+    def remove_view(self, view: BaseView) -> None:
         for item in view.walk_children():
             if item.is_storable():
                 self._views.pop((item.type.value, item.custom_id), None)  # type: ignore
@@ -1012,10 +1328,15 @@ class ViewStore:
                 self.remove_message_view(key)
                 break
 
-    def remove_message_view(self, message_id):
+    def remove_message_view(self, message_id: int) -> None:
         del self._synced_message_views[message_id]
 
-    def dispatch(self, component_type: int, custom_id: str, interaction: Interaction):
+    def get_message_view(self, message_id: int) -> BaseView | None:
+        return self._synced_message_views.get(message_id)
+
+    def dispatch(
+        self, component_type: int, custom_id: str, interaction: Interaction
+    ) -> None:
         self.__verify_integrity()
         message_id: int | None = interaction.message and interaction.message.id
         key = (component_type, message_id, custom_id)
@@ -1032,13 +1353,15 @@ class ViewStore:
         item.refresh_state(interaction)
         view._dispatch_item(item, interaction)
 
-    def is_message_tracked(self, message_id: int):
+    def is_message_tracked(self, message_id: int) -> bool:
         return message_id in self._synced_message_views
 
     def remove_message_tracking(self, message_id: int) -> BaseView | None:
         return self._synced_message_views.pop(message_id, None)
 
-    def update_from_message(self, message_id: int, components: list[ComponentPayload]):
+    def update_from_message(
+        self, message_id: int, components: list[ComponentPayload]
+    ) -> None:
         # pre-req: is_message_tracked == true
         view = self._synced_message_views[message_id]
         components = [_component_factory(d, state=self._state) for d in components]